deleted:    .ipynb_checkpoints/fig2a-checkpoint.ipynb 	deleted:    .ipynb_checkpoints/fig2b-checkpoint.ipynb 	deleted:    .ipynb_checkpoints/ibmq-AQT-checkpoint.ipynb 	deleted:    .ipynb_checkpoints/ibmq-getdata-checkpoint.ipynb 	deleted:    .ipynb_checkpoints/sim-AQT-large-checkpoint.ipynb 	modified:   __pycache__/povm.cpython-36.pyc 	new file:   data/aqt-ghz-large-400_2-64-4_6-130000_pt.npy 	new file:   data/aqt-ghz-large-400_2-64-4_6-200000_pt.npy 	new file:   data/aqt-ghz-large-400_2-64-4_6-400000_pt.npy 	new file:   data/aqt-ghz-large-400_2-64-4_6-40000_pt.npy 	new file:   data/aqt-ghz-large-400_2-64-4_6-80000_pt.npy 	new file:   data/aqt-ghz-large-600_2-64-4_6-40000_pt.npy 	new file:   data/ghz-sim-Pauli6-Nq6-Ns130000.npy 	new file:   data/ghz-sim-Pauli6-Nq6-Ns200000.npy 	new file:   data/ghz-sim-Pauli6-Nq6-Ns40000.npy 	new file:   data/ghz-sim-Pauli6-Nq6-Ns400000.npy 	new file:   data/ghz-sim-Pauli6-Nq6-Ns80000.npy 	new file:   error_scaling.ipynb 	new file:   errorscaling.py 	deleted:    figures/Fig2_NP.pdf 	deleted:    figures/Fig2_NP.xcf 	deleted:    figures/Fig3_NP.pdf 	deleted:    figures/Fig3_NP.xcf 	deleted:    figures/Fig4_NP.pdf 	deleted:    figures/fig2a.pdf 	deleted:    figures/fig2b.pdf 	deleted:    figures/ibmq-AQT.pdf 	deleted:    figures/ibmq-MLE.pdf 	new file:   mod/aqt-ghz-large-400_2-64-4_6-130000.mod 	new file:   mod/aqt-ghz-large-400_2-64-4_6-40000.mod 	new file:   mod/aqt-ghz-large-400_2-64-4_6-400000.mod 	new file:   mod/aqt-ghz-large-400_2-64-4_6-80000.mod 	new file:   mod/aqt-ghz-large-600_2-64-4_6-40000.mod 	modified:   povm.py 	modified:   sim-AQT-large.ipynb

+++ povm.py
@@ -342,10 +342,11 @@ def LoadData(Nq, fname):
 
 
 def GenerateData(gen, Ns, fname):
-
-    print('Generating data: %s'%fname)
-    
-    data = gen.samples(Ns=Ns)
-
-    np.save(fname, data)
+    if os.path.isfile(fname):
+        print('Loading data: %s'%fname)
+        data = np.load(fname)
+    else:
+        print('Generating data: %s'%fname)
+        data = gen.samples(Ns=Ns)
+        np.save(fname, data)
     return data
